chore: refactor code with card components
